project04/requirements.txt
gitpython>=3.1.0
colorama>=0.4.4
